WEBAPP-358: Rename pointsOfInterest to poi and change route to locations
