+++ src/modules/endpoint/reducers/index.js
@@ -8,7 +8,7 @@ import disclaimerEndpoint from '../endpoints/disclaimer'
 import extrasEndpoint from '../endpoints/extras'
 import sprungbrettJobEndpoint from '../endpoints/sprungbrettJobs'
 import wohnenEndpoint from '../endpoints/wohnen'
-import pointsOfInterestEndpoint from '../endpoints/pointsOfInterest'
+import pointsOfInterestEndpoint from '../endpoints/pois'
 
 import { handleActions } from 'redux-actions'
 import Payload from '../Payload'
